add reponse error

+++ src/modules/customer/useCases/deleteCustomer/DeleteCustomerUseCase.ts
@@ -1,7 +1,6 @@
 import { inject, injectable } from 'tsyringe';
 
 import { IUCustomersRepository } from '../../repositories/ICustomersRepository';
-import { Customer } from '../../entities/Customer';
 
 @injectable()
 class DeleteCustomerUseCase {
@@ -10,8 +9,18 @@ class DeleteCustomerUseCase {
     private customerRepository: IUCustomersRepository
   ) {}
 
-  async execute(id: string): Promise<Customer> {
-    return await this.customerRepository.delete(id);
+  async execute(id: string): Promise<any> {
+    const idNull = !!id === false;
+    console.log('idnull', idNull);
+    if (idNull) {
+      return { error: 'Id is required!', status: 400 };
+    }
+    console.log(id);
+    try {
+      return await this.customerRepository.delete(id);
+    } catch (error) {
+      return { error: 'Not found', status: 404 };
+    }
   }
 }
